Move sketchfab zip to worker
